test: adding threats to threat model object
I think it makes sense to require an identifier (either str or UUID -
str such that if the threat modeler has some predefined IDs we can
pass them in, else a UUID will get generated) for __getitem__
and __contains__. If they want to use the Element and Threat objects
themselves to check for membership in the ThreatModel the developer
should access the lists directly, i.e. threat in ThreatModel.threats.