test, app: pure refactor - store elements, threats as dict
This removes some logic for __getitem__ and __contains__ that
basically reproduces what a dict is doing, just way less efficiently.
Now each element in ThreatModel.elements and ThreatModel.threats is
an ID: Class pair, where ID can be a str or UUID, and Class is the
corresponding object type for the collection.